Modifications

+++ app/Http/Controllers/Api/V1/AdminController.php
@@ -2,16 +2,17 @@
 
 namespace App\Http\Controllers\Api\V1;
 
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Gate;
-use App\Models\Api\V1\Admin;
-use App\Models\Api\V1\Category;
-use Illuminate\Support\Facades\Auth;
-
 use App\Models\Api\V1\Role;
 use App\Models\Api\V1\User;
+use App\Models\Api\V1\Admin;
+use Illuminate\Http\Request;
+use App\Models\Api\V1\Category;
+use App\Models\Api\V1\Permission;
+
 use Illuminate\Support\Facades\DB;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\Api\V1\CreateAdminRequest;
 use App\Http\Requests\Api\V1\UpdateAdminRequest;
 
@@ -31,6 +32,46 @@ class AdminController extends Controller
         } catch (\Exception $e) {
             return response()->json([
                 'status' => 0,
+            ]);
+        }
+    }
+
+    // Index permissions list for admin
+    public function indexAdminPermissions()
+    {
+        try {
+            Gate::authorize('isSuperAdmin');
+            $role = Role::where('name', 'admin')->first();
+            $permissions = Permission::where('role_id', $role->id)->get();
+
+            return response()->json([
+                'status' => 1,
+                'permissions' => $permissions,
+            ]);
+        } catch (\Exception $e) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'An error occurred while retrieving admin permissions.',
+            ]);
+        }
+    }
+
+    // Index permissions list for provider
+    public function indexProviderPermissions()
+    {
+        try {
+            Gate::authorize('isSuperAdmin');
+            $role = Role::where('name', 'provider')->first();
+            $permissions = Permission::where('role_id', $role->id)->get();
+
+            return response()->json([
+                'status' => 1,
+                'permissions' => $permissions,
+            ]);
+        } catch (\Exception $e) {
+            return response()->json([
+                'status' => 0,
+                'message' => 'An error occurred while retrieving provider permissions.',
             ]);
         }
     }
@@ -81,16 +122,27 @@ class AdminController extends Controller
                 $admin = Admin::where('admin_id', $id)->first();
                 $user = User::find($id);
                 $user->update(['role' => 'patient']);
-                $admin->delete();
+
+                // Check if the admin exists to handle the null case
+                if ($admin) {
+                    $admin->delete();
+                } else {
+                    return response()->json([
+                        'status' => 0,
+                        'message' => 'Admin does not exist.',
+                    ]);
+                }
+
                 return response()->json([
                     'status' => 1,
+                    'message' => 'Admin deleted successfully.',
                 ]);
             });
         } catch (\Exception $e) {
             return response()->json([
                 'status' => 0,
+                'message' => 'An error occurred while deleting the admin.',
             ]);
         }
-
     }
 }
